Model/ModelGCP/phmodels/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
pandas
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
openai[embeddings]
h5py
gunicorn==20.1.0
beautifulsoup4==4.12.2
nltk